@@ -17,8 +17,8 @@ package sk.antons.web.filter.util;
 
 import java.io.IOException;
 import java.io.InputStream;
-import javax.servlet.ReadListener;
-import javax.servlet.ServletInputStream;
+import jakarta.servlet.ReadListener;
+import jakarta.servlet.ServletInputStream;
 
 /**
  * Simple ServletInputStream implementation.
