@@ -17,8 +17,8 @@ package sk.antons.web.filter.util;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import javax.servlet.ServletOutputStream;
-import javax.servlet.WriteListener;
+import jakarta.servlet.ServletOutputStream;
+import jakarta.servlet.WriteListener;
 
 /**
  * Simple ServletOutputStream implementation. 
